+++ app/containers/AlbumPage/AlbumPage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Helmet } from 'react-helmet';
-
+// import albumData from '../../setup/data/data';
 /* 
 * since theres no state to pass, 
 * will leave as stateless function
